Move `AsyncDictionary` code to its own file.
